More fix for the code

+++ 14-binary_tree_balance.c
@@ -8,8 +8,8 @@
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t left_height;
-	size_t right_height;
+	int left_height;
+	int right_height;
 
 	if (tree)
 	{
